多光谱segformer模型配置完成

+++ voc_annotation.py
@@ -64,7 +64,8 @@ if __name__ == "__main__":
 
     print("Check datasets format, this may take a while.")
     print("检查数据集格式是否符合要求，这可能需要一段时间。")
-    classes_nums        = np.zeros([256], np.int)
+    # np.int 在新版 NumPy 中已经移除；这里用内置 int 保持兼容。
+    classes_nums        = np.zeros([256], int)
     for i in tqdm(list):
         name            = total_seg[i]
         png_file_name   = os.path.join(segfilepath, name)
